# models/credential_metadata.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"

	"github.com/go-openapi/swag"
)

// CredentialMetadata credential metadata
// swagger:model CredentialMetadata
type CredentialMetadata struct {

	// The block number where the transaction was inserted
	BlockNumber string `json:"blockNumber,omitempty"`

	// The dateTime when the block was generated YYYY-MM-DDTHH:MM:SSZ
	Timestamp string `json:"timestamp,omitempty"`

	// The transaction hash
	Transaction string `json:"transaction,omitempty"`

	// mail of contact
	Email string `json:"email,omitempty"`
}

// Validate validates this credential metadata
func (m *CredentialMetadata) Validate(formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *CredentialMetadata) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *CredentialMetadata) UnmarshalBinary(b []byte) error {
	var res CredentialMetadata
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}